refactor(personal-assistant): extract 11 leaf domains to sub-services
Phase B (batch 1): extract the base-only / simple-connector mixins into
domains/*-service.ts sub-services taking LifeOpsContext; mixins now thin
delegators preserving byte-identical public signatures. Domains: google,
calendar, relationships, health, imessage, telegram, signal, whatsapp, x-read
(injects a resolveXGrant callback for the optional sibling grant resolver),
email-unsubscribe, subscriptions.

Plugin typechecks clean (0 errors); test suite unchanged (105 pass / 4 pre-
existing fails identical to baseline, issue #8234). No behaviour change.
