+++ plugins/plugin-personal-assistant/src/lifeops/service-mixin-email-unsubscribe.ts
@@ -1,4 +1,3 @@
-import { InboxUnsubscribeService } from "@elizaos/plugin-inbox/inbox/unsubscribe-service";
 import type {
   EmailSubscriptionScanResult,
   EmailUnsubscribeRecord,
@@ -6,6 +5,7 @@ import type {
   EmailUnsubscribeResult,
   EmailUnsubscribeScanRequest,
 } from "@elizaos/plugin-inbox/inbox/email-unsubscribe-types";
+import { EmailUnsubscribeDomain } from "./domains/email-unsubscribe-service.js";
 import type {
   Constructor,
   LifeOpsServiceBase,
@@ -26,15 +26,9 @@ export interface LifeOpsEmailUnsubscribeService {
 }
 
 /**
- * Email-unsubscribe back-end moved to `@elizaos/plugin-inbox`
- * ({@link InboxUnsubscribeService}). This mixin is now a thin delegation layer:
- * it preserves the LifeOpsService method surface (the `requestUrl` argument the
- * route callers pass), but forwards to the standalone inbox service, which
- * resolves Gmail through the `@elizaos/plugin-google` runtime service and
- * persists to the same `app_inbox.life_email_unsubscribes` table.
- *
- * The two-phase confirmation gate (`requireConfirmation`) stays in the PA route
- * layer; the inbox service trusts the pre-confirmed `userAuthorization` flag.
+ * Thin delegation layer over {@link EmailUnsubscribeDomain}. Preserves the
+ * LifeOpsService method surface (including the `requestUrl` argument the route
+ * callers pass) and forwards to the domain sub-service.
  *
  * @internal
  */
@@ -42,34 +36,38 @@ export function withEmailUnsubscribe<
   TBase extends Constructor<LifeOpsServiceBase>,
 >(Base: TBase): MixinClass<TBase, LifeOpsEmailUnsubscribeService> {
   class LifeOpsEmailUnsubscribeMixin extends Base {
-    private inboxUnsubscribeService(): InboxUnsubscribeService {
-      return new InboxUnsubscribeService(this.runtime);
-    }
+    // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+    // Public (not private) to avoid TS4094 on the re-exported mixin class.
+    readonly emailUnsubscribeDomain = new EmailUnsubscribeDomain(this);
 
     async scanEmailSubscriptions(
-      _requestUrl: URL,
+      requestUrl: URL,
       request: EmailUnsubscribeScanRequest = {},
     ): Promise<EmailSubscriptionScanResult> {
-      return this.inboxUnsubscribeService().scanEmailSubscriptions(request);
+      return this.emailUnsubscribeDomain.scanEmailSubscriptions(
+        requestUrl,
+        request,
+      );
     }
 
     async unsubscribeEmailSender(
-      _requestUrl: URL,
+      requestUrl: URL,
       request: EmailUnsubscribeRequest,
     ): Promise<EmailUnsubscribeResult> {
-      return this.inboxUnsubscribeService().unsubscribeEmailSender(request);
+      return this.emailUnsubscribeDomain.unsubscribeEmailSender(
+        requestUrl,
+        request,
+      );
     }
 
     async listEmailUnsubscribes(
       limit = 100,
     ): Promise<EmailUnsubscribeRecord[]> {
-      return this.inboxUnsubscribeService().listEmailUnsubscribes(limit);
+      return this.emailUnsubscribeDomain.listEmailUnsubscribes(limit);
     }
 
     summarizeEmailUnsubscribeScan(result: EmailSubscriptionScanResult): string {
-      return this.inboxUnsubscribeService().summarizeEmailUnsubscribeScan(
-        result,
-      );
+      return this.emailUnsubscribeDomain.summarizeEmailUnsubscribeScan(result);
     }
   }
