+++ plugins/plugin-personal-assistant/src/lifeops/service-mixin-google.ts
@@ -1,7 +1,3 @@
-import {
-  type ConnectorAccount,
-  getConnectorAccountManager,
-} from "@elizaos/core";
 import type {
   DisconnectLifeOpsGoogleConnectorRequest,
   LifeOpsConnectorGrant,
@@ -11,28 +7,12 @@ import type {
   StartLifeOpsGoogleConnectorRequest,
   StartLifeOpsGoogleConnectorResponse,
 } from "../contracts/index.js";
-import {
-  disconnectedGoogleStatus,
-  googleAccountIdFromGrantId,
-  googleGrantFromAccount,
-  googleGrantIdForAccount,
-  googleScopesForAccount,
-  googleSideForAccount,
-  googleStatusFromAccount,
-  listGoogleConnectorAccounts,
-  resolveGoogleConnectorAccount,
-} from "./google-plugin-delegates.js";
+import { GoogleDomain } from "./domains/google-service.js";
 import type {
   Constructor,
   LifeOpsServiceBase,
   MixinClass,
 } from "./service-mixin-core.js";
-import { fail, normalizeOptionalString } from "./service-normalize.js";
-import {
-  normalizeGoogleCapabilityRequest,
-  normalizeOptionalConnectorMode,
-  normalizeOptionalConnectorSide,
-} from "./service-normalize-connector.js";
 
 export interface LifeOpsGoogleService {
   getGoogleConnectorStatus(
@@ -63,64 +43,6 @@ export interface LifeOpsGoogleService {
   ): Promise<LifeOpsGoogleConnectorStatus>;
 }
 
-function roleForSide(side: LifeOpsConnectorSide): "OWNER" | "AGENT" {
-  return side === "agent" ? "AGENT" : "OWNER";
-}
-
-function sideFromMetadata(value: unknown): LifeOpsConnectorSide {
-  return value === "agent" ? "agent" : "owner";
-}
-
-function requestedScopesForCapabilities(
-  capabilities: readonly string[] | undefined,
-): string[] | undefined {
-  if (!capabilities || capabilities.length === 0) {
-    return undefined;
-  }
-  return googleScopesForAccount(
-    {
-      id: "requested",
-      provider: "google",
-      role: "OWNER",
-      purpose: [],
-      accessGate: "owner_binding",
-      status: "pending",
-      createdAt: Date.now(),
-      updatedAt: Date.now(),
-      metadata: { grantedCapabilities: [...capabilities] },
-    } as ConnectorAccount,
-    capabilities as never,
-  );
-}
-
-function assertLocalMode(mode?: LifeOpsConnectorMode): void {
-  if (mode && mode !== "local") {
-    fail(
-      410,
-      "LifeOps no longer manages cloud or legacy Google modes. Use @elizaos/plugin-google connector accounts.",
-    );
-  }
-}
-
-function googlePluginUnavailableStatus(
-  side: LifeOpsConnectorSide,
-): LifeOpsGoogleConnectorStatus {
-  return {
-    ...disconnectedGoogleStatus(side),
-    configured: false,
-    reason: "config_missing",
-    degradations: [
-      {
-        axis: "disconnected",
-        code: "google_plugin_unavailable",
-        message:
-          "@elizaos/plugin-google is required for Google accounts. LifeOps no longer stores Google OAuth tokens directly.",
-        retryable: true,
-      },
-    ],
-  };
-}
-
 /** @internal */
 export function withGoogle<TBase extends Constructor<LifeOpsServiceBase>>(
   Base: TBase,
@@ -128,415 +50,157 @@ export function withGoogle<TBase extends Constructor<LifeOpsServiceBase>>(
   const GoogleBase = Base;
 
   class LifeOpsGoogleServiceMixin extends GoogleBase {
-    private googleConnectorManager() {
-      try {
-        return getConnectorAccountManager(this.runtime);
-      } catch {
-        return null;
-      }
-    }
+    // `this` (a LifeOpsServiceBase subclass) satisfies LifeOpsContext.
+    // Public (not private) to avoid TS4094 on the re-exported mixin class.
+    readonly googleDomain = new GoogleDomain(this);
 
-    private googleAccountStatus(
-      account: ConnectorAccount,
-    ): LifeOpsGoogleConnectorStatus {
-      return googleStatusFromAccount({
-        account,
-        agentId: this.agentId(),
-        defaultMode: "local",
-        availableModes: ["local"],
-      });
-    }
-
-    public async withGoogleGrantOperation<T>(
+    public withGoogleGrantOperation<T>(
       _grant: LifeOpsConnectorGrant,
       operation: () => Promise<T>,
     ): Promise<T> {
-      return operation();
+      return this.googleDomain.withGoogleGrantOperation(_grant, operation);
     }
 
-    public async runManagedGoogleOperation<T>(
+    public runManagedGoogleOperation<T>(
       _grant: LifeOpsConnectorGrant,
       _operation: () => Promise<T>,
     ): Promise<T> {
-      fail(
-        410,
-        "Cloud-managed Google operations were removed from LifeOps. Use @elizaos/plugin-google connector accounts.",
-      );
+      return this.googleDomain.runManagedGoogleOperation(_grant, _operation);
     }
 
-    public async clearGoogleConnectorData(
+    public clearGoogleConnectorData(
       side?: LifeOpsConnectorSide,
     ): Promise<void> {
-      const calendarEvents = await this.repository.listCalendarEvents(
-        this.agentId(),
-        "google",
-        undefined,
-        undefined,
-        side,
-      );
-      await this.deleteCalendarReminderPlansForEvents(
-        calendarEvents.map((event) => event.id),
-      );
-      await this.repository.deleteCalendarEventsForProvider(
-        this.agentId(),
-        "google",
-        undefined,
-        side,
-      );
-      await this.repository.deleteCalendarSyncState(
-        this.agentId(),
-        "google",
-        undefined,
-        side,
-      );
-      await this.repository.deleteGmailMessagesForProvider(
-        this.agentId(),
-        "google",
-        side,
-      );
-      await this.repository.deleteGmailSpamReviewItemsForProvider(
-        this.agentId(),
-        "google",
-        side,
-      );
-      await this.repository.deleteGmailSyncState(
-        this.agentId(),
-        "google",
-        undefined,
-        side,
-      );
+      return this.googleDomain.clearGoogleConnectorData(side);
     }
 
-    public async clearGoogleGrantData(
-      grant: LifeOpsConnectorGrant,
-    ): Promise<void> {
-      await this.deleteCalendarReminderPlansForEvents(
-        (
-          await this.repository.listCalendarEvents(
-            this.agentId(),
-            "google",
-            undefined,
-            undefined,
-            grant.side,
-          )
-        )
-          .filter((event) => event.grantId === grant.id)
-          .map((event) => event.id),
-      );
-      await this.repository.deleteCalendarEventsForProvider(
-        this.agentId(),
-        "google",
-        grant.id,
-        grant.side,
-      );
-      await this.repository.deleteCalendarSyncState(
-        this.agentId(),
-        "google",
-        grant.id,
-        grant.side,
-      );
-      await this.repository.deleteGmailMessagesForProvider(
-        this.agentId(),
-        "google",
-        grant.side,
-        grant.id,
-      );
-      await this.repository.deleteGmailSpamReviewItemsForProvider(
-        this.agentId(),
-        "google",
-        grant.side,
-        grant.id,
-      );
-      await this.repository.deleteGmailSyncState(
-        this.agentId(),
-        "google",
-        grant.id,
-        grant.side,
-      );
+    public clearGoogleGrantData(grant: LifeOpsConnectorGrant): Promise<void> {
+      return this.googleDomain.clearGoogleGrantData(grant);
     }
 
-    public async deleteCalendarReminderPlansForEvents(
+    public deleteCalendarReminderPlansForEvents(
       _eventIds: string[],
     ): Promise<void> {
-      // Implemented by withCalendar; this no-op fallback keeps withGoogle
-      // independently usable in unit tests that compose only connector status
-      // methods.
+      return this.googleDomain.deleteCalendarReminderPlansForEvents(_eventIds);
     }
 
-    public async setPreferredGoogleConnectorMode(
+    public setPreferredGoogleConnectorMode(
       _mode: LifeOpsConnectorMode | null,
       _side?: LifeOpsConnectorSide,
     ): Promise<LifeOpsConnectorGrant | null> {
-      return null;
+      return this.googleDomain.setPreferredGoogleConnectorMode(_mode, _side);
     }
 
-    public async requireGoogleCalendarGrant(
+    public requireGoogleCalendarGrant(
       requestUrl: URL,
       requestedMode?: LifeOpsConnectorMode,
       requestedSide?: LifeOpsConnectorSide,
       grantId?: string,
     ): Promise<LifeOpsConnectorGrant> {
-      assertLocalMode(normalizeOptionalConnectorMode(requestedMode, "mode"));
-      const status = await this.getGoogleConnectorStatus(
-        requestUrl,
-        "local",
-        requestedSide,
-        grantId,
-      );
-      const grant = status.grant;
-      if (!status.connected || !grant) {
-        fail(409, "Google Calendar is not connected.");
-      }
-      if (!grant.capabilities.includes("google.calendar.read")) {
-        fail(403, "Google Calendar read access has not been granted.");
-      }
-      return grant;
-    }
-
-    public async requireGoogleCalendarWriteGrant(
-      requestUrl: URL,
-      requestedMode?: LifeOpsConnectorMode,
-      requestedSide?: LifeOpsConnectorSide,
-      grantId?: string,
-    ): Promise<LifeOpsConnectorGrant> {
-      const grant = await this.requireGoogleCalendarGrant(
+      return this.googleDomain.requireGoogleCalendarGrant(
         requestUrl,
         requestedMode,
         requestedSide,
         grantId,
       );
-      if (!grant.capabilities.includes("google.calendar.write")) {
-        fail(403, "Google Calendar write access has not been granted.");
-      }
-      return grant;
     }
 
-    public async requireGoogleGmailGrant(
+    public requireGoogleCalendarWriteGrant(
       requestUrl: URL,
       requestedMode?: LifeOpsConnectorMode,
       requestedSide?: LifeOpsConnectorSide,
       grantId?: string,
     ): Promise<LifeOpsConnectorGrant> {
-      assertLocalMode(normalizeOptionalConnectorMode(requestedMode, "mode"));
-      const status = await this.getGoogleConnectorStatus(
-        requestUrl,
-        "local",
-        requestedSide,
-        grantId,
-      );
-      const grant = status.grant;
-      if (!status.connected || !grant) {
-        fail(409, "Google Gmail is not connected.");
-      }
-      if (!grant.capabilities.includes("google.gmail.triage")) {
-        fail(403, "Google Gmail triage access has not been granted.");
-      }
-      return grant;
-    }
-
-    public async requireGoogleGmailSendGrant(
-      requestUrl: URL,
-      requestedMode?: LifeOpsConnectorMode,
-      requestedSide?: LifeOpsConnectorSide,
-      grantId?: string,
-    ): Promise<LifeOpsConnectorGrant> {
-      const grant = await this.requireGoogleGmailGrant(
+      return this.googleDomain.requireGoogleCalendarWriteGrant(
         requestUrl,
         requestedMode,
         requestedSide,
         grantId,
       );
-      if (!grant.capabilities.includes("google.gmail.send")) {
-        fail(403, "Google Gmail send access has not been granted.");
-      }
-      return grant;
     }
 
-    async getGoogleConnectorStatus(
-      _requestUrl: URL,
+    public requireGoogleGmailGrant(
+      requestUrl: URL,
+      requestedMode?: LifeOpsConnectorMode,
+      requestedSide?: LifeOpsConnectorSide,
+      grantId?: string,
+    ): Promise<LifeOpsConnectorGrant> {
+      return this.googleDomain.requireGoogleGmailGrant(
+        requestUrl,
+        requestedMode,
+        requestedSide,
+        grantId,
+      );
+    }
+
+    public requireGoogleGmailSendGrant(
+      requestUrl: URL,
+      requestedMode?: LifeOpsConnectorMode,
+      requestedSide?: LifeOpsConnectorSide,
+      grantId?: string,
+    ): Promise<LifeOpsConnectorGrant> {
+      return this.googleDomain.requireGoogleGmailSendGrant(
+        requestUrl,
+        requestedMode,
+        requestedSide,
+        grantId,
+      );
+    }
+
+    getGoogleConnectorStatus(
+      requestUrl: URL,
       requestedMode?: LifeOpsConnectorMode,
       requestedSide?: LifeOpsConnectorSide,
       grantId?: string,
     ): Promise<LifeOpsGoogleConnectorStatus> {
-      const mode = normalizeOptionalConnectorMode(requestedMode, "mode");
-      assertLocalMode(mode);
-      const side =
-        normalizeOptionalConnectorSide(requestedSide, "side") ?? "owner";
-      const manager = this.googleConnectorManager();
-      if (!manager?.getProvider?.("google")) {
-        return googlePluginUnavailableStatus(side);
-      }
-      const account = await resolveGoogleConnectorAccount({
-        runtime: this.runtime,
-        requestedSide: side,
+      return this.googleDomain.getGoogleConnectorStatus(
+        requestUrl,
+        requestedMode,
+        requestedSide,
         grantId,
-      });
-      return account
-        ? this.googleAccountStatus(account)
-        : disconnectedGoogleStatus(side);
+      );
     }
 
-    async getGoogleConnectorAccounts(
-      _requestUrl: URL,
+    getGoogleConnectorAccounts(
+      requestUrl: URL,
       requestedSide?: LifeOpsConnectorSide,
     ): Promise<LifeOpsGoogleConnectorStatus[]> {
-      const side = normalizeOptionalConnectorSide(requestedSide, "side");
-      const manager = this.googleConnectorManager();
-      if (!manager?.getProvider?.("google")) {
-        return side
-          ? [googlePluginUnavailableStatus(side)]
-          : [
-              googlePluginUnavailableStatus("owner"),
-              googlePluginUnavailableStatus("agent"),
-            ];
-      }
-      const accounts = await listGoogleConnectorAccounts({
-        runtime: this.runtime,
-        requestedSide: side,
-      });
-      if (accounts.length === 0) {
-        return side ? [disconnectedGoogleStatus(side)] : [];
-      }
-      return accounts.map((account) => this.googleAccountStatus(account));
+      return this.googleDomain.getGoogleConnectorAccounts(
+        requestUrl,
+        requestedSide,
+      );
     }
 
-    async selectGoogleConnectorMode(
+    selectGoogleConnectorMode(
       requestUrl: URL,
       preferredModeInput: LifeOpsConnectorMode | undefined,
       requestedSide?: LifeOpsConnectorSide,
     ): Promise<LifeOpsGoogleConnectorStatus> {
-      const preferredMode = normalizeOptionalConnectorMode(
+      return this.googleDomain.selectGoogleConnectorMode(
+        requestUrl,
         preferredModeInput,
-        "mode",
+        requestedSide,
       );
-      assertLocalMode(preferredMode);
-      return this.getGoogleConnectorStatus(requestUrl, "local", requestedSide);
     }
 
-    async startGoogleConnector(
+    startGoogleConnector(
       request: StartLifeOpsGoogleConnectorRequest,
       requestUrl: URL,
     ): Promise<StartLifeOpsGoogleConnectorResponse> {
-      const mode = normalizeOptionalConnectorMode(request.mode, "mode");
-      assertLocalMode(mode);
-      const requestedSide =
-        normalizeOptionalConnectorSide(request.side, "side") ?? "owner";
-      const requestedCapabilities = normalizeGoogleCapabilityRequest(
-        request.capabilities,
-      );
-      const manager = this.googleConnectorManager();
-      if (!manager?.getProvider?.("google")) {
-        fail(
-          503,
-          "@elizaos/plugin-google is required before starting Google OAuth.",
-        );
-      }
-
-      const requestedAccountId = googleAccountIdFromGrantId(request.grantId);
-      const redirectUri = new URL(
-        "/api/connectors/google/oauth/callback",
-        requestUrl.origin,
-      ).toString();
-      const flow = await manager.startOAuth("google", {
-        redirectUri,
-        accountId: requestedAccountId ?? undefined,
-        scopes: requestedScopesForCapabilities(requestedCapabilities),
-        metadata: {
-          lifeops: true,
-          side: requestedSide,
-          role: roleForSide(requestedSide),
-          requestedRole: roleForSide(requestedSide),
-          requestedCapabilities,
-          privacy: "owner_only",
-          redirectUrl: normalizeOptionalString(request.redirectUrl),
-        },
-      });
-      return {
-        provider: "google",
-        side: requestedSide,
-        mode: "local",
-        requestedCapabilities: requestedCapabilities ?? [],
-        redirectUri: flow.redirectUri ?? redirectUri,
-        authUrl: flow.authUrl ?? "",
-      };
+      return this.googleDomain.startGoogleConnector(request, requestUrl);
     }
 
-    async completeGoogleConnectorCallback(
+    completeGoogleConnectorCallback(
       callbackUrl: URL,
     ): Promise<LifeOpsGoogleConnectorStatus> {
-      const manager = this.googleConnectorManager();
-      if (!manager?.getProvider?.("google")) {
-        fail(
-          503,
-          "@elizaos/plugin-google is required before completing Google OAuth.",
-        );
-      }
-      const state = callbackUrl.searchParams.get("state") ?? "";
-      const completed = await manager.completeOAuth("google", {
-        state,
-        code: callbackUrl.searchParams.get("code") ?? undefined,
-        error: callbackUrl.searchParams.get("error") ?? undefined,
-        errorDescription:
-          callbackUrl.searchParams.get("error_description") ?? undefined,
-        query: Object.fromEntries(callbackUrl.searchParams.entries()),
-      });
-      const side = sideFromMetadata(completed.flow.metadata?.side);
-      const accountId =
-        completed.account?.id ?? completed.flow.accountId ?? undefined;
-      return this.getGoogleConnectorStatus(
-        callbackUrl,
-        "local",
-        side,
-        accountId ? googleGrantIdForAccount(accountId) : undefined,
-      );
+      return this.googleDomain.completeGoogleConnectorCallback(callbackUrl);
     }
 
-    async disconnectGoogleConnector(
+    disconnectGoogleConnector(
       request: DisconnectLifeOpsGoogleConnectorRequest,
       requestUrl: URL,
     ): Promise<LifeOpsGoogleConnectorStatus> {
-      const mode = normalizeOptionalConnectorMode(request.mode, "mode");
-      assertLocalMode(mode);
-      const side = normalizeOptionalConnectorSide(request.side, "side");
-      const manager = this.googleConnectorManager();
-      if (!manager?.getProvider?.("google")) {
-        fail(
-          503,
-          "@elizaos/plugin-google is required before disconnecting Google accounts.",
-        );
-      }
-      const requestedGrantId = normalizeOptionalString(request.grantId);
-      const account = await resolveGoogleConnectorAccount({
-        runtime: this.runtime,
-        requestedSide: side,
-        grantId: requestedGrantId,
-      });
-      if (!account) {
-        if (requestedGrantId) {
-          fail(404, "Google connector account not found.");
-        }
-        return this.getGoogleConnectorStatus(requestUrl, "local", side);
-      }
-      const grant = googleGrantFromAccount({
-        account,
-        agentId: this.agentId(),
-      });
-      await manager.deleteAccount("google", account.id);
-      await this.clearGoogleGrantData(grant);
-      await this.recordConnectorAudit(
-        "google:connector-account",
-        "google connector account disconnected",
-        {
-          connectorAccountId: account.id,
-          side: googleSideForAccount(account),
-        },
-        { disconnected: true },
-      );
-      return this.getGoogleConnectorStatus(
-        requestUrl,
-        "local",
-        side ?? googleSideForAccount(account),
-      );
+      return this.googleDomain.disconnectGoogleConnector(request, requestUrl);
     }
   }
